remove global cache and use on-demand request-level caching
