+++ server/storage/disk/fragment_impl.go
@@ -34,7 +34,6 @@ import (
 	"github.com/zilliztech/woodpecker/common/metrics"
 	"github.com/zilliztech/woodpecker/common/werr"
 	"github.com/zilliztech/woodpecker/server/storage"
-	"github.com/zilliztech/woodpecker/server/storage/cache"
 )
 
 var _ storage.Fragment = (*FragmentFileWriter)(nil)
@@ -515,17 +514,24 @@ type FragmentFileReader struct {
 	fragmentId int64 // Unique identifier for the fragment
 	filePath   string
 	fileSize   int64
+
+	// data
 	mappedFile mmap.MMap
 	fd         *os.File
 
+	// info
 	firstEntryID int64 // ID of the first entry
 	lastEntryID  int64 // ID of the last entry
 	entryCount   int32 // Current number of entries
 	isGrowing    bool  // Whether this fragment will continue to receive writes
 
+	// status
 	infoFetched bool
 	dataLoaded  bool
 	closed      bool // Whether the file is closed
+
+	// data refCnt
+	dataRefCnt int // The number of references to the fragment data used
 }
 
 // NewFragmentFileReader creates a new FragmentFile, which can read only
@@ -546,6 +552,8 @@ func NewFragmentFileReader(filePath string, fileSize int64, logId int64, segment
 		infoFetched: false,
 		dataLoaded:  false,
 		closed:      false,
+
+		dataRefCnt: 0, // mmap not open, initial reference count is zero
 	}
 	return ff, nil
 }
@@ -577,10 +585,16 @@ func (fr *FragmentFileReader) Flush(ctx context.Context) error {
 func (fr *FragmentFileReader) Load(ctx context.Context) error {
 	fr.mu.Lock()
 	defer fr.mu.Unlock()
+	// already loaded, no need to load again
+	if fr.dataLoaded {
+		// Increase the reference count
+		fr.dataRefCnt += 1
+		return nil
+	}
+
 	start := time.Now()
 	logId := fmt.Sprintf("%d", fr.logId)
 	segmentId := fmt.Sprintf("%d", fr.segmentId)
-
 	if fr.closed {
 		return errors.New("fragment file is closed")
 	}
@@ -620,14 +634,10 @@ func (fr *FragmentFileReader) Load(ctx context.Context) error {
 	fr.infoFetched = true
 	fr.dataLoaded = true
 
+	// Increase the reference count
+	fr.dataRefCnt += 1
+
 	logger.Ctx(ctx).Debug("fragment file loaded", zap.Int64("fragmentId", fr.fragmentId), zap.String("filePath", fr.filePath), zap.String("fragInst", fmt.Sprintf("%p", fr)))
-
-	// update cache
-	err = cache.AddCacheFragment(ctx, fr)
-	if err != nil {
-		logger.Ctx(ctx).Warn("add fragment to cache failed ", zap.String("fragmentPath", fr.filePath), zap.Int64("fragmentId", fr.fragmentId), zap.Error(err))
-	}
-
 	// update metrics
 	metrics.WpFragmentLoadBytes.WithLabelValues(logId, segmentId).Add(float64(fr.fileSize))
 	metrics.WpFragmentLoadTotal.WithLabelValues(logId, segmentId).Inc()
@@ -697,8 +707,8 @@ func (fr *FragmentFileReader) readFooter() error {
 	return nil
 }
 
-// check and release resource immediately, avoid memory surges when catching up read
-func (fr *FragmentFileReader) isMMapReadable(ctx context.Context) bool {
+// IsMMapReadable check and release resource immediately, avoid memory surges when catching up read
+func (fr *FragmentFileReader) IsMMapReadable(ctx context.Context) bool {
 	fr.mu.Lock()
 	defer fr.mu.Unlock()
 
@@ -811,69 +821,57 @@ func (fr *FragmentFileReader) refreshFooter() error {
 // GetLastEntryId returns the last entry ID.
 func (fr *FragmentFileReader) GetLastEntryId() (int64, error) {
 	fr.mu.RLock()
+	defer fr.mu.RUnlock()
 	if fr.closed {
-		fr.mu.RUnlock()
 		return 0, errors.New("fragment file is closed")
 	}
+	// if infoFetched and not growing, return
+	if fr.infoFetched && !fr.isGrowing {
+		// already fetch and not growing, return
+		lastEntryID := fr.lastEntryID
+		return lastEntryID, nil
+	}
 
-	if fr.infoFetched && fr.dataLoaded && fr.isGrowing {
-		// already fetch but it is growing, refresh
-		fr.mu.RUnlock()
-		fr.mu.Lock()
-		defer fr.mu.Unlock()
+	// otherwise it is growing or info not fetched, all these cases need to load data first
+	if fr.dataLoaded {
 		err := fr.refreshFooter()
 		if err != nil {
 			return -1, err
 		}
-		return fr.lastEntryID, nil
-	} else if fr.infoFetched && !fr.isGrowing {
-		// already fetch and not growing, return
 		lastEntryID := fr.lastEntryID
-		fr.mu.RUnlock()
-		return lastEntryID, nil
-	} else {
-		// Need to load first
-		fr.mu.RUnlock()
-		// No need to lock before Load, as Load already acquires a write lock
-		err := fr.Load(context.Background())
-		if err != nil {
-			return -1, err
-		}
-
-		// Need to get the result with read lock
-		fr.mu.RLock()
-		lastEntryID := fr.lastEntryID
-		fr.mu.RUnlock()
 		return lastEntryID, nil
 	}
+
+	// need to fetch info first manually
+	return -1, werr.ErrFragmentInfoNotFetched.WithCauseErrMsg(fmt.Sprintf("%s info not fetched", fr.filePath))
 }
 
 // GetFirstEntryId returns the first entry ID.
 func (fr *FragmentFileReader) GetFirstEntryId() (int64, error) {
 	fr.mu.RLock()
+	defer fr.mu.RUnlock()
 	if fr.closed {
-		fr.mu.RUnlock()
 		return 0, errors.New("fragment file is closed")
 	}
-
-	if fr.infoFetched {
+	// if infoFetched and not growing, return
+	if fr.infoFetched && !fr.isGrowing {
+		// already fetch and not growing, return
 		firstEntryID := fr.firstEntryID
-		fr.mu.RUnlock()
 		return firstEntryID, nil
 	}
-	fr.mu.RUnlock()
 
-	// No need to lock before Load, as Load already acquires a write lock
-	err := fr.Load(context.Background())
-	if err != nil {
-		return 0, err
+	// otherwise it is growing or info not fetched, all these cases need to load data first
+	if fr.dataLoaded {
+		err := fr.refreshFooter()
+		if err != nil {
+			return -1, err
+		}
+		firstEntryID := fr.firstEntryID
+		return firstEntryID, nil
 	}
 
-	// Need to get the result with read lock
-	fr.mu.RLock()
-	firstEntryID := fr.firstEntryID
-	fr.mu.RUnlock()
-	return firstEntryID, nil
+	// need to fetch info first manually
+	return -1, werr.ErrFragmentInfoNotFetched.WithCauseErrMsg(fmt.Sprintf("%s info not fetched", fr.filePath))
 }
 
 // GetLastModified returns the last modification time.
@@ -888,13 +886,16 @@ func (fr *FragmentFileReader) GetLastModified() int64 {
 // GetEntry returns the entry at the specified ID.
 func (fr *FragmentFileReader) GetEntry(entryId int64) ([]byte, error) {
 	fr.mu.RLock()
+	defer fr.mu.RUnlock()
 	if fr.closed {
 		logger.Ctx(context.Background()).Warn("failed to get entry from a closed fragment file",
 			zap.String("filePath", fr.filePath),
 			zap.Int64("fragmentId", fr.fragmentId),
 			zap.Int64("readingEntryId", entryId))
-		fr.mu.RUnlock()
 		return nil, errors.New("fragment file is closed")
+	}
+	if !fr.dataLoaded {
+		return nil, werr.ErrFragmentNotLoaded.WithCauseErrMsg(fmt.Sprintf("%s not loaded", fr.filePath))
 	}
 
 	logger.Ctx(context.Background()).Debug("Try get entry from this fragment",
@@ -904,127 +905,7 @@ func (fr *FragmentFileReader) GetEntry(entryId int64) ([]byte, error) {
 		zap.Int64("readingEntryId", entryId),
 		zap.String("fragInst", fmt.Sprintf("%p", fr)))
 
-	// Load data if not loaded
-	if !fr.dataLoaded {
-		fr.mu.RUnlock()
-		// No need to lock before Load, as Load already acquires a write lock
-		err := fr.Load(context.Background())
-		if err != nil {
-			return nil, err
-		}
-
-		// Now check entry range with read lock
-		fr.mu.RLock()
-		// Refresh footer info only if entry is beyond the last known entry
-		if entryId > fr.lastEntryID && fr.isGrowing {
-			fr.mu.RUnlock()
-			fr.mu.Lock()
-			defer fr.mu.Unlock()
-
-			err := fr.refreshFooter()
-			if err != nil {
-				return nil, err
-			}
-
-			// Check if entryId is within range after refresh
-			if entryId < fr.firstEntryID || entryId > fr.lastEntryID {
-				logger.Ctx(context.Background()).Debug("entry ID out of range after refresh",
-					zap.Int64("requestedID", entryId),
-					zap.Int64("firstEntryID", fr.firstEntryID),
-					zap.Int64("lastEntryID", fr.lastEntryID))
-				return nil, werr.ErrInvalidEntryId.WithCauseErrMsg(fmt.Sprintf("entry ID %d not in the range of this fragment", entryId))
-			}
-
-			return fr.getEntryLocked(entryId)
-		}
-
-		// Check if entryId is within range - either with read lock or write lock depending on path taken
-		if entryId < fr.firstEntryID || entryId > fr.lastEntryID {
-			logger.Ctx(context.Background()).Debug("entry ID out of range",
-				zap.Int64("requestedID", entryId),
-				zap.Int64("firstEntryID", fr.firstEntryID),
-				zap.Int64("lastEntryID", fr.lastEntryID))
-			fr.mu.RUnlock()
-			return nil, werr.ErrInvalidEntryId.WithCauseErrMsg(fmt.Sprintf("entry ID %d not in the range of this fragment", entryId))
-		}
-	}
-
-	// Copy required data with read lock
-	mappedFile := fr.mappedFile
-	fileSize := fr.fileSize
-	filePath := fr.filePath
-	firstEntryID := fr.firstEntryID
-	defer fr.mu.RUnlock()
-
-	// Calculate index position - relative position of entry ID in index area
-	idxPos := uint32(fileSize - footerSize - int64(indexItemSize)*(int64(entryId-firstEntryID+1)))
-
-	if idxPos < headerSize || idxPos >= uint32(fileSize-footerSize) {
-		logger.Ctx(context.Background()).Debug("Invalid index position",
-			zap.Uint32("idxPos", idxPos),
-			zap.Uint32("headerSize", headerSize),
-			zap.Int64("fileSize", fileSize),
-			zap.Uint32("footerSize", footerSize))
-		return nil, fmt.Errorf("invalid index position: %d", idxPos)
-	}
-
-	// Read data offset
-	offset := binary.LittleEndian.Uint32(mappedFile[idxPos:])
-	if offset < headerSize || offset >= uint32(fileSize-footerSize) {
-		logger.Ctx(context.Background()).Debug("Invalid data offset",
-			zap.Uint32("offset", offset),
-			zap.Uint32("headerSize", headerSize),
-			zap.Int64("fileSize", fileSize))
-		return nil, fmt.Errorf("invalid data offset: %d", offset)
-	}
-
-	// Read data length
-	length := binary.LittleEndian.Uint32(mappedFile[offset:])
-	if length == 0 || length > uint32(fileSize-footerSize)-offset-8 {
-		logger.Ctx(context.Background()).Debug("Invalid data length",
-			zap.Uint32("length", length),
-			zap.Uint32("offset", offset),
-			zap.Int64("fileSize", fileSize))
-		return nil, fmt.Errorf("invalid data length: %d", length)
-	}
-
-	// Read CRC (4 bytes)
-	storedCRC := binary.LittleEndian.Uint32(mappedFile[offset+4:])
-
-	// Determine data region
-	dataStart := offset + 8 // Skip length(4 bytes) and CRC(4 bytes)
-	dataEnd := dataStart + length
-	if dataEnd > uint32(fileSize-footerSize) {
-		logger.Ctx(context.Background()).Debug("Data region out of bounds",
-			zap.Uint32("dataStart", dataStart),
-			zap.Uint32("dataEnd", dataEnd),
-			zap.Int64("fileSize", fileSize))
-		return nil, fmt.Errorf("data region out of bounds: %d-%d", dataStart, dataEnd)
-	}
-
-	// Read data
-	data := make([]byte, length)
-	copy(data, mappedFile[dataStart:dataEnd])
-
-	logger.Ctx(context.Background()).Debug("Fragment data read completed",
-		zap.String("fragmentFile", filePath),
-		zap.Int64("readingEntryId", entryId),
-		zap.Uint32("start", dataStart),
-		zap.Uint32("end", dataEnd),
-		zap.Uint32("idxPos", idxPos),
-		zap.Int("dataSize", len(data)),
-		zap.String("fragInst", fmt.Sprintf("%p", fr)))
-
-	// Verify CRC
-	if crc32.ChecksumIEEE(data) != storedCRC {
-		logger.Ctx(context.Background()).Debug("CRC mismatch",
-			zap.Int64("entryId", entryId),
-			zap.Uint32("computedCRC", crc32.ChecksumIEEE(data)),
-			zap.Uint32("storedCRC", storedCRC))
-		return nil, fmt.Errorf("CRC mismatch for entry ID %d", entryId)
-	}
-
-	return data, nil
+	return fr.getEntryLocked(entryId)
 }
 
 // getEntryLocked is a helper method to retrieve an entry with lock already held
@@ -1216,10 +1097,18 @@ func (fr *FragmentFileReader) Release() error {
 		return nil
 	}
 
+	// decrement reference count
+	fr.dataRefCnt -= 1
+
+	if fr.dataRefCnt > 0 {
+		// still in use, no need to release
+		return nil
+	}
+
 	// Unmap memory mapping
 	if fr.mappedFile != nil {
 		if err := fr.mappedFile.Unmap(); err != nil {
-			return errors.Wrap(err, "failed to unmap fragment file")
+			logger.Ctx(context.Background()).Warn("failed to unmap fragment file", zap.String("filePath", fr.filePath))
 		}
 		fr.mappedFile = nil
 	}
@@ -1229,6 +1118,7 @@ func (fr *FragmentFileReader) Release() error {
 		if err := fr.fd.Close(); err != nil {
 			logger.Ctx(context.Background()).Warn("failed to close fragment file", zap.String("filePath", fr.filePath))
 		}
+		fr.fd = nil
 	}
 
 	// Mark data as not fetched in buffer
